pyproject.toml
[tool.poetry]
name = "twelete"
version = "0.1.0"
description = "Automatically delete tweets"
authors = ["Sander van Dragt <sander@vandragt.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "^3.9.0"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
